webkit/data/layout_test_results/v8/LayoutTests/fast/js/toString-prefix-postfix-preserve-parens-expected.txt

CONSOLE MESSAGE: line 7: Uncaught SyntaxError: Invalid left-hand side expression in postfix operation
FAIL successfullyParsed should be true. Threw exception ReferenceError: successfullyParsed is not defined

TEST COMPLETE
